+++ CHANGELOG.md
@@ -4,6 +4,8 @@ change log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [1.4.0.32]
 ### Added
 - Reagent interop for Tab
 
@@ -56,7 +58,8 @@ Update clojars with more repository information.
 ### Added
 - Reagent interop with @headlessui/react for Disclosure, Transition and FocusTrap
 
-[Unreleased]: https://github.com/mainej/headlessui-reagent/compare/v1.2.1...main
+[Unreleased]: https://github.com/mainej/headlessui-reagent/compare/v1.4.0.32...main
+[1.4.0.32]: https://github.com/mainej/headlessui-reagent/compare/v1.2.1...v1.4.0.32
 [1.2.1]: https://github.com/mainej/headlessui-reagent/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/mainej/headlessui-reagent/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/mainej/headlessui-reagent/compare/v1.0.0...v1.1.0